+++ src/org/omnifaces/showcase/utils/AjaxBean.java
@@ -38,12 +38,12 @@ public class AjaxBean {
 		Ajax.oncomplete("showData()");
 	}
 
-	public void updateRow(UIData table, long index) { // Due to JBoss 7.0 bug, we can't use int argument. This is fixed in 7.1.
-		Ajax.updateRow(table, (int) index);
+	public void updateRow(UIData table, int index) {
+		Ajax.updateRow(table, index);
 	}
 
-	public void updateColumn(UIData table, long index) { // Due to JBoss 7.0 bug, we can't use int argument. This is fixed in 7.1.
-		Ajax.updateColumn(table, (int) index);
+	public void updateColumn(UIData table, int index) {
+		Ajax.updateColumn(table, index);
 	}
 
 }
